+++ src/LucidMongo/QueryBuilder/index.js
@@ -12,7 +12,7 @@
 const _ = require('lodash')
 const query = require('mquery')
 const debug = require('debug')('mquery')
-
+// const GeoPoint = require('geo-point')
 const EagerLoad = require('../EagerLoad')
 // const RelationsParser = require('../Relations/Parser')
 const CE = require('../../Exceptions')
@@ -542,18 +542,12 @@ class QueryBuilder {
       this._eagerLoads[args[0]] = args[1]
     } else if (args[1] && _.isObject(args[1])) {
       this._eagerLoads[args[0]] = (builder) => {
-        _.forEach(args[1], (value, key) => {
-          builder[key](value)
-        })
+        _.forEach(args[1], (value, key) => builder[key](value))
       }
     } else if (Array.isArray(args[0])) {
-      _.forEach(args[0], related => {
-        this.with(related)
-      })
+      _.forEach(args[0], related => this.with(related))
     } else if (_.isObject(args[0])) {
-      _.forEach(args[0], (scope, key) => {
-        this.with(key, scope)
-      })
+      _.forEach(args[0], (scope, key) => this.with(key, scope))
     } else {
       this._eagerLoads[args[0]] = (builder) => { }
     }
@@ -647,6 +641,13 @@ class QueryBuilder {
       'in',
       'nin',
       'all',
+      'near',
+      'maxDistance',
+      'mod',
+      'includes',
+      'polygon',
+      'elemMatch',
+      'geometry',
       'intersects'
     ]
   }
@@ -669,38 +670,6 @@ class QueryBuilder {
   }
 
   /**
-   * Support Methods
-   *
-   * @readonly
-   * @static
-   * @memberof QueryBuilder
-   */
-  static get supportMethods () {
-    return [
-      'all',
-      'exists',
-      'elemMatch',
-      'eq',
-      'gt',
-      'gte',
-      'lt',
-      'lte',
-      'in',
-      'ne',
-      'nin',
-      'nor',
-      'regex',
-      'size',
-      'mod',
-      'slice',
-      'intersects',
-      'regex',
-      'maxDistance',
-      'minDistance'
-    ]
-  }
-
-  /**
    * Replace where method
    *
    * @returns {this}
@@ -708,50 +677,34 @@ class QueryBuilder {
    */
   where () {
     if (_.isPlainObject(arguments[0])) {
-      _.forEach(arguments[0], (conditions, key) => {
-        if (key === 'and' || key === 'or' || key === 'nor') {
+      let queryObject = arguments[0]
+      for (const key in queryObject) {
+        const conditions = queryObject[key]
+        if (key === '$and' || key === '$or' || key === '$nor') {
           if (!Array.isArray(conditions)) {
             throw new CE.InvalidArgumentException(`Method "$${key}"'s param must be an array`)
           }
-          let queries = []
-          _.forEach(conditions, (condition) => {
-            queries.push(query().where(condition)._conditions)
-          })
-          this.query[key](queries)
+          let formatedConditions = []
+          for (const condition of conditions) {
+            formatedConditions.push(query().where(condition)._conditions)
+          }
+          queryObject[key] = formatedConditions
         } else if (_.isPlainObject(conditions)) {
-          _.forEach(conditions, (c, k) => {
-            if (k === 'near' || k === 'nearSphere') {
-              let point = { center: [c.lng, c.lat] }
-              if (this.Model.geometries && this.Model.geometries.includes(key)) {
-                point = { center: { type: 'Point', coordinates: [c.lng, c.lat], spherical: k === 'nearSphere' } }
-              }
-              if (conditions.maxDistance) {
-                point.maxDistance = conditions.maxDistance
-              }
-              if (conditions.minDistance) {
-                point.minDistance = conditions.minDistance
-              }
-              this.query.where(key).near(point)
-            } else if (this.constructor.supportMethods.includes(k)) {
-              if (k !== 'maxDistance' && k !== 'minDistance') {
-                this.query.where(key)[k](c)
-              }
-            } else {
-              throw new CE.RuntimeException(`Method "$${k}" is not support by query builder`)
+          for (const subKey in conditions) {
+            const reg = /^\$(eq|ne|gt|gte|lt|lte|in|nin|all|near|intersects|elemMatch|includes)$/
+            if (reg.test(subKey)) {
+              queryObject[key][subKey] = this.Model.formatField(key, queryObject[key][subKey])
             }
-          })
-        } else {
-          const value = this.Model.formatField(key, conditions)
-          this.query.where(key, value)
+          }
         }
-      })
+        queryObject[key] = this.Model.formatField(key, queryObject[key])
+      }
+      this.query.where(queryObject)
     } else if (_.isFunction(arguments[0])) {
       arguments[0].bind(this).call()
     } else {
       if (arguments.length === 2) {
-        const key = arguments[0]
-        const value = this.Model.formatField(arguments[0], arguments[1])
-        this.query.where(key, value)
+        this.query.where(arguments[0]).eq(arguments[1])
       } else if (arguments.length === 3) {
         switch (arguments[1]) {
           case '=':
@@ -932,7 +885,7 @@ class QueryBuilder {
   async _aggregate (aggregator, key, groupBy) {
     this._applyScopes()
     const $match = this.query._conditions
-    const $group = { }
+    const $group = {}
     if (_.isString(groupBy)) {
       $group._id = '$' + groupBy
     } else if (_.isObject) {