+++ src/LucidMongo/Serializers/LucidMongoSerializer.js
@@ -113,7 +113,7 @@ class LucidMongoSerializer {
    */
   async findById (id) {
     debug('finding user with primary key as %s', id)
-    const user = this._getQuery().where(this.primaryKey, id).first()
+    const user = await this._getQuery().where(this.primaryKey, id).first()
     if (user) {
       user.$attributes[this.primaryKey] = String(user.primaryKeyValue)
       return user
@@ -168,7 +168,12 @@ class LucidMongoSerializer {
    * @return {Object|Null}
    */
   async findByToken (token, type) {
+    if (!this._Token) {
+      return null
+    }
+
     debug('finding user for %s token', token)
+
     token = await this._Token.where({ token, type, is_revoked: false }).first()
     if (token) {
       const user = await token.user().first()